pyproject.toml
[project]
name = "ages-backend"
version = "0.0.1"
description = "Mapa digital backend code"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "aioboto3>=11.0",
    "asyncpg>=0.30",
    "dotenv>=0.9.9",
    "slowapi>=0.1.9",
    "email-validator>=2.1",
    "fastapi>=0.135.1",
    "bcrypt>=4.0",
    "pydantic-settings>=2.13.1",
    "PyJWT>=2.8",
    "sqlalchemy[asyncio]>=2.0",
    "uvicorn>=0.41.0",
    "python-multipart>=0.0.9",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]
extend-exclude = ["script_integration_tests.py"]

[tool.ruff.lint]
select = ["E","W","F","I","B","C4","UP","D"]
ignore = ["E203","D104","B008"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["D100","D104","F401"]
"test_*.py" = ["D100","D101","D102","F401"]
"tests/**/*.py" = ["D100","D101","D102","F401"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "aiosqlite>=0.20",
    "coverage>=7.13.1",
    "httpx>=0.28",
    "pyright>=1.1",
    "pytest>=8.0",
    "ruff>=0.14.13",
]

[tool.pyright]
include = ["src", "tests"]
exclude = [
    "**/__pycache__",
    "**/.venv",
    "**/htmlcov",
    "script_integration_tests.py",
]
extraPaths = ["src"]
pythonVersion = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
concurrency = ["greenlet", "thread"]
omit = [
    "src/__init__.py",
    "src/md_backend/__init__.py",
    "src/md_backend/models/__init__.py",
    "src/md_backend/routes/__init__.py",
    "src/md_backend/services/__init__.py",
    "src/md_backend/utils/__init__.py",
]

[tool.coverage.report]
fail_under = 100
show_missing = true
